requirements.txt
streamlit
pandas
yfinance==0.2.41
pyxirr
nsepython
requests
python-telegram-bot
schedule
python-dotenv